Travis builds: always build manylinux wheels
This also reformats the YAML, for which I apologize, but it was hard to
work on it without doing that.

It drops dist: xenial which has become the default.

It renames matrix: to jobs: because that's the preferres spelling
according to Travis CI docs.

It adds names to all the special build jobs.

It drops Python 3.4 builds on Mac OS.

It *doesn't* add Python 3.8 builds for Mac OS because AFAIU terryfy
doesn't support 3.8 yet.

It moves docker pull and twine upload out of .manylinux.sh and into
.travis.yml.

It unifies twine upload calls by supplifying username/password via
environment variables.

The changes largely mirror those in
https://github.com/zopefoundation/BTrees/pull/115.