Make sure all test files are used
Rename test_pickle_{2,3}.py because these are not supposed to be
collected automatically: they use syntax incompatible with Python 3 and
2 respectively.

Get rid of the -m zodbpickle.tests.test_pickle$ hack that was used to
avoid the aforementioned syntax errors, and was also preventing the new
test_compile_flags.py from being executed.

Get rid of unittest.makeSuite() because I hear it's getting deprecated.